% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{activation.fork}
\alias{activation.fork}
\title{Estimate optimum of expression and time of activation}
\usage{
activation.fork(r, fork.de, mat, root, leaves, deriv.cutoff = 0.015,
  gamma = 1, n.mapping = 1, n.cores = 10)
}
\arguments{
\item{r}{ppt.tree object}

\item{fork.de}{outcome of test.fork.genes function}

\item{mat}{expression matrix}

\item{root}{root of progenitor branch of bifurcation}

\item{leaves}{leaves of derivative branches of bifurcation}

\item{deriv.cutoff}{a first passage of derivative through cutoff 'deriv.cutoff' to predict activation timing}

\item{gamma}{gamma parameter in gam function}

\item{n.mapping}{results are averaged among n.mapping number of probabilsitic cell projections}

\item{n.cores}{number of cores to use}
}
\value{
table fork.de with added per gene timing of optimum and activation
}
\description{
Estimate optimum of expression and time of activation
}
