% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{activation.statistics}
\alias{activation.statistics}
\title{Estimate optimum of expression and time of activation}
\usage{
activation.statistics(r, mat, root, leave, genes = rownames(mat),
  deriv.cutoff = 0.015, gamma = 1, n.mapping = 1,
  n.cores = parallel::detectCores()/2)
}
\arguments{
\item{r}{ppt.tree object}

\item{mat}{expression matrix}

\item{root}{root of progenitor branch of bifurcation}

\item{genes}{genes to estimate parameters}

\item{deriv.cutoff}{a first passage of derivative through cutoff 'deriv.cutoff' to predict activation timing}

\item{gamma}{gamma parameter in gam function}

\item{n.mapping}{results are averaged among n.mapping number of probabilsitic cell projections}

\item{n.cores}{number of cores to use}

\item{leaves}{leaves of derivative branches of bifurcation}
}
\value{
per gene timing of optimum and activation
}
\description{
Estimate optimum of expression and time of activation
}
