% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{activity.lasso}
\alias{activity.lasso}
\title{Predict regulatory impact (activity) of transcription factors}
\usage{
activity.lasso(em, motmat, perm = FALSE, n.cores = 1)
}
\arguments{
\item{em}{matrix of expression levels}

\item{motmat}{matrix of target-TF scores}

\item{perm}{boolean, do permutations if TRUE.}

\item{n.cores}{number of cores to use}
}
\value{
matrix of predited TF activities in cells.
}
\description{
Predict regulatory impact (activity) of transcription factors
}
