% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{bootstrap.ppt}
\alias{bootstrap.ppt}
\title{Sample pptree objects using bootstrap}
\usage{
bootstrap.ppt(..., X, M = ncol(X), n.cores = 20, n.samples = n.cores,
  seed = NULL, replace = T)
}
\arguments{
\item{X}{expression matrix of genes (rows) and cells (columns).}

\item{M}{number of principal points of pptree.}

\item{n.samples}{a number of seed samplings.}

\item{replace}{sampling with replacement (logical).}
}
\value{
a list of pptree objects
}
\description{
Sample pptree objects using bootstrap
}
