% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{classify.genes}
\alias{classify.genes}
\title{Classify tree-associated genes}
\usage{
classify.genes(r, n.cores = 20, cutoff = 0.2)
}
\arguments{
\item{r}{tree}

\item{cutoff}{expression in local optimum should be higher/lower than both terminal branch values by cutoff.}

\item{X}{expressinon matrix of genes (rows) vs cell (columns)}
}
\value{
vector of predicted classification for fitted genes.
}
\description{
Tree-associated genes are classified in branch-monotonous, transiently expressed and having complex patterns.
}
