% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{cor.mat}
\alias{cor.mat}
\title{Calculate weighted pairwise correlations between columns of matrices A and B}
\usage{
cor.mat(A, B)
}
\description{
Calculate weighted pairwise correlations between columns of matrices A and B
}
