% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{corMatrixC}
\alias{corMatrixC}
\title{Estimates local gene-gene correlations in a sequence of windows along pseudotime for locally permuted matrix}
\usage{
corMatrixC(genes, cells, cdnormX, step, w, permutN,
  n.cores = parallel::detectCores()/2)
}
\arguments{
\item{genes}{a set of genes to estimate local pairwise expression correlations}

\item{cells}{cells ordered by pseudotime}

\item{cdnormX}{expression matrix}

\item{step}{step between sequential windows (in number of cells)}

\item{w}{window length along pseudotime (in number of cells)}

\item{permutN}{number of local permutations}

\item{n.cores}{number of cores to use}
}
\value{
local multiple permuted gene-gene correlation matrices per each window of cells
}
\description{
Estimates local gene-gene correlations in a sequence of windows along pseudotime for locally permuted matrix
}
