% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crest.R
\docType{data}
\name{crest}
\alias{crest}
\title{Processed neural crest single-cell data}
\format{A list of 6 items.
\describe{
  \item{fpm}{normalized expression matrix}
  \item{wgm}{matrix of expression levels adjusted for mean-variance trend}
  \item{wgwm}{matrix of expression weights}
  \item{emb}{2D t-SNE embedding}
  \item{clcol}{cell colors of clusters}
\item{nc.cells}{a vector of neural crest cells. Other cells are neural tube cells.}
}}
\source{
\url{web_link}
}
\usage{
data(crest)
}
\description{
A dataset containing normalized expression matrices, t-SNE embedding and clusters.
}
\keyword{datasets}
