% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{euclidean.mat}
\alias{euclidean.mat}
\title{Calculate pairwise euclidean distances between columns of matrices A and B}
\usage{
euclidean.mat(A, B)
}
\description{
Calculate pairwise euclidean distances between columns of matrices A and B
}
