% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{extract.subtree}
\alias{extract.subtree}
\title{Extract subtree of the tree}
\usage{
extract.subtree(r, nodes)
}
\arguments{
\item{r}{ppt.tree object}

\item{nodes}{set tips or internal nodes (bifurcations) to extract subtree}
}
\value{
list of segments comprising a subtree.
}
\description{
Extract subtree of the tree
}
