% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{fig.cor.evol}
\alias{fig.cor.evol}
\title{Estimate probabil}
\usage{
fig.cor.evol(cors, freq, emb, genesetA, genesetB)
}
\arguments{
\item{n.samples}{a number of seed samplings.}

\item{seeds}{a vector of seeds to use. Overwrites n.samples.}
}
\value{
a list of pptree objects
}
\description{
Estimate probabil
}
