% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{fig.cors}
\alias{fig.cors}
\title{Estimate probabil}
\usage{
fig.cors(cors, genesetA, genesetB, val = FALSE)
}
\arguments{
\item{n.samples}{a number of seed samplings.}

\item{seeds}{a vector of seeds to use. Overwrites n.samples.}
}
\value{
a list of pptree objects
}
\description{
Estimate probabil
}
