% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{fit.ts}
\alias{fit.ts}
\title{Model gene expression levels as a brancing spline function of tree positions.}
\usage{
fit.ts(r, X, n.map, n.cores = parallel::detectCores()/2, gamma = 1.5,
  knn = 1)
}
\arguments{
\item{r}{pptree object}

\item{X}{expressinon matrix of genes (rows) vs cells (columns)}

\item{n.map}{number of probabilistic cell-to-tree mappings to use}

\item{gamma}{stringency of penalty.}

\item{knn}{use expression averaging among knn cells}
}
\value{
matrix of fitted gene expression levels to the tree
}
\description{
Model gene expression levels as a brancing spline function of tree positions.
}
