% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{inclusion.stat}
\alias{inclusion.stat}
\title{Show summary cumulative probability of genes inclusion in correlated module}
\usage{
inclusion.stat(alps, incl.real, incl.perm)
}
\arguments{
\item{alps}{levels of alp parameter}

\item{incl.real}{average inclusion time for alp levels}

\item{incl.perm}{average inclusion time for alp levels in permuted matrix of expression levels}
}
\description{
Show summary cumulative probability of genes inclusion in correlated module
}
