% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{lambda.explore}
\alias{lambda.explore}
\title{Explore lambda}
\usage{
lambda.explore(X = NA, M = ncol(X), sigma = 0.1, emb = NA,
  metrics = "cosine", tips.min = 2, tips.max = 20, base = 2,
  lambda.init = 100, err.cut = 0.005, n.steps = 40, p.power = 1)
}
\arguments{
\item{X}{gene (rows) vs cell (columns) expression matrix}

\item{M}{number of principal points in pptree modeling}

\item{sigma}{fixed parameter sigma used in pptree modeling}

\item{emb}{embdedding to visalize cells and principal tree together. If emb is given than pptrees for a range of lambda are shown}
}
\description{
Explores multiple lambda and choose the optimal
}
