% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{loc.opt}
\alias{loc.opt}
\title{Identify all local optima for a time series data}
\usage{
loc.opt(series, cutoff = 0.1)
}
\arguments{
\item{series}{- time series data}

\item{cutoff}{- expression in local optimum should be on cutoff higher/lower than nearby local optima. This parameter allows to eliminate small local optimas that are likely artifacts}
}
\value{
data frame containing type of local optima (min/max) and time index.
}
\description{
Identify all local optima for a time series data
}
