% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matWVCorr}
\alias{matWVCorr}
\title{calculate weighted correlation between columns of a matrix and a given vector}
\usage{
matWVCorr(Mat, Vec, Vecw)
}
\description{
calculate weighted correlation between columns of a matrix and a given vector
}
