% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{mppt.tree}
\alias{mppt.tree}
\title{Sample pptree objects using different seeds}
\usage{
mppt.tree(..., n.cores = parallel::detectCores()/2,
  n.samples = n.cores, seed = NULL, seeds = NULL)
}
\arguments{
\item{n.samples}{a number of seed samplings.}

\item{seeds}{a vector of seeds to use. Overwrites n.samples.}
}
\value{
a list of pptree objects
}
\description{
Sample pptree objects using different seeds
}
