% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{onset.est}
\alias{onset.est}
\title{Estimates inclusion times of genes in a correlated module}
\usage{
onset.est(ppt, mapping = 1, geneset, nodes, expr, alp = 20, w = 40,
  step = 10, do.perm = FALSE, winperm = 30, n.cells = 280,
  permut.n = 10, n.cores = 1, autocor.cut = 0.95, iterations = 15)
}
\arguments{
\item{ppt}{ppt tree}

\item{mapping}{id of probabilistic cells project to use}

\item{geneset}{a set of genes that form a module}

\item{nodes}{tips of a tree that form subtree (e.g, trajectory or fork)}

\item{expr}{expression matrix}

\item{alp}{parameter regulating stringency of inclusion event}

\item{w}{local window of cells along pseudotime to estimate local correlation}

\item{step}{shift of a window along pseudotime in number of cells}

\item{do.perm}{do local estimates for locally permuted expression matrix}

\item{winperm}{number of local cells for permutations}

\item{n.cells}{total number of cells to consider from root in progenitor branch}

\item{permut.n}{of local cells for permutations}

\item{n.cores}{number of cores to calculate permutations used to estimate background local correlations in onset.est}

\item{autocor.cut}{cutoff on correlation of inclusion times between sequential iterations of the algorithm to stop it}

\item{iterations}{maximum number of iterations of the algorithm}
}
\value{
vector of inclusion times of genes
}
\description{
Estimates inclusion times of genes in a correlated module
}
