% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{onset}
\alias{onset}
\title{Estimates inclusion times of genes in a correlated module for a number of probabilistic cell projections}
\usage{
onset(ppt, geneset, nodes = NULL, expr, alp = 20, w = 40,
  step = 10, n.cells = 280, mappings = 1, do.perm = FALSE,
  winperm = w, n.cores = parallel::detectCores()/2, permut.n = 10,
  n.cores1 = 1)
}
\arguments{
\item{ppt}{ppt tree}

\item{geneset}{a set of genes that form a module}

\item{nodes}{tips of a tree that form subtree (e.g, trajectory or fork)}

\item{expr}{expression matrix}

\item{alp}{parameter regulating stringency of inclusion event}

\item{w}{local window of cells along pseudotime to estimate local correlation}

\item{step}{shift of a window along pseudotime in number of cells}

\item{n.cells}{total number of cells to consider from root in progenitor branch}

\item{mappings}{number of probabilistic cell projections to assess}

\item{do.perm}{do local estimates for locally permuted expression matrix}

\item{winperm}{number of local cells for permutations}

\item{n.cores}{number of cores}

\item{permut.n}{number of permutations used to estiamte background local correlations in onset.est}

\item{n.cores1}{number of cores to calculate permutations used to estimate background local correlations in onset.est}
}
\value{
matrix of inclusion timing for each gene (rows) in each probabilistic cells projection (columns)
}
\description{
Estimates inclusion times of genes in a correlated module for a number of probabilistic cell projections
}
