% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{plotppt}
\alias{plotppt}
\title{Visualize pptree onto embedding}
\usage{
plotppt(r, emb, F = NULL, gene = NULL, main = gene, mat = NULL,
  pattern.cell = NULL, pattern.tree = NULL, cex.col = NA,
  tree.col = NULL, cex.main = 0.5, cex.title = 1, cex.tree = 1.5,
  lwd.tree = 1, par = TRUE, tips = FALSE, forks = FALSE,
  subtree = NA, ...)
}
\arguments{
\item{r}{- pptree object}

\item{emb}{- (x,y) coordinates data frame (e.g Rtsne $Y result)}

\item{F}{- coordinates of principal points (optional)}

\item{gene}{- a gene to show expression of (optional)}

\item{mat}{- gene vs cell expression matrix (needed if option 'gene' is activated)}

\item{pattern.cell}{- numeric profile of a quantity for each cell (e.g. expression of a gene or cell cycle stage)}

\item{pattern.tree}{- numeric profile of a quantity for each principal point (e.g. expression of a gene or cell cycle stage)}

\item{cex.col}{- color of points}

\item{cex.main}{- cex of points}

\item{cex.title}{- cex of title}

\item{cex.tree}{- cex of principal points}

\item{tips}{- logical, to draw indecies of tips of the tree. Usefull before usage of cleanup.branches()}
}
\description{
Projects pptree onto embedding (e.g. tSNE)
}
