% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{plotpptl}
\alias{plotpptl}
\title{Visualize list of pptree objects onto embedding}
\usage{
plotpptl(rl, emb, cols = adjustcolor(1, alpha = 0.3), alpha = 1,
  lwd = 1, ...)
}
\arguments{
\item{rl}{list of pptree objects (as calculated using bootstrap.tree or mppt.tree)}

\item{emb}{(x,y) coordinates data frame (e.g Rtsne $Y result)}

\item{cols}{vector of colors for cells in emb.}
}
\description{
Projects pptree objects onto embedding (e.g. tSNE)
}
