% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{ppt.tree}
\alias{ppt.tree}
\title{Reconstruction of the tree}
\usage{
ppt.tree(X, W = NA, emb = NA, M, init = NULL, plot = TRUE,
  output = TRUE, lambda = 10, sigma = 0.1, seed = NULL,
  n.steps = 50, err.cut = 0.05, metrics = "cosine", p.power = 1,
  knn = NULL, ...)
}
\arguments{
\item{X}{gene (row) vs cell (columns) expression matrix}

\item{emb}{embdedding to visalize cells and principal tree together}

\item{M}{number of principal points to use (more than zero, no more than number of cells)}

\item{init}{matrix of initial gene coordinates of principal points}

\item{plot}{plot or not intermediate trees}

\item{lambda}{penalty for the tree  length, as used in SimplePPT}

\item{sigma}{parameter as used in SimplePPT}

\item{seed}{used to make initial assignment of principal points to a subset of cells}

\item{n.steps}{number of iteraions}

\item{err.cut}{stop algorithm if proximity of principal points between iterations less than err.cut}

\item{metrics}{metrics used to calculated distances between cells or principal points. "euclidean" or "cosine"}

\item{p.power}{if cosine metrics used, option p.power allows to use (1-cor)^p.power (p.power=1 by default)}
}
\value{
pptree object
}
\description{
Using SimplePPT approach to model principal tree (pptree) of the data
}
