% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{project.cells.onto.ppt}
\alias{project.cells.onto.ppt}
\title{Project cells onto the principal tree}
\usage{
project.cells.onto.ppt(r, emb = NULL, n.mapping = 1)
}
\arguments{
\item{r}{pptree object}

\item{emb}{if not NULL than cell branch assignment and color code of branches are shown}

\item{n.mapping}{number of probabilistic mapping of cells onto the tree to use. If n.mapping=1 then likelihood cell mapping is used.}
}
\value{
modified pptree object with new fields r$cell.summary, r$cell.info and r$img.list. r$cell.summary contains information about cells projected onto the tree, including pseudotime and branch.
}
\description{
Project cells onto the principal tree
}
