% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{setroot}
\alias{setroot}
\title{Orient the tree by setting up the root}
\usage{
setroot(r, root = NULL, plot = TRUE)
}
\arguments{
\item{r}{pptree object}

\item{root}{root principal point (plotppt(tips=TRUE,..) can be used to visualize candidate tips for a root)}
}
\value{
modified ppt.tree object with new fields r$pp.info (estimated pseudotime and branch of principal points), r$pp.segments (segments information), r$root (root id).
}
\description{
Assign root, pseudotime and segment to each principal point of the tree
}
