% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{show.gene.inclusion}
\alias{show.gene.inclusion}
\title{Show cumulative probability of gene inclusion in correlated module}
\usage{
show.gene.inclusion(gn, matSwitch, bins = 8, par = FALSE)
}
\arguments{
\item{gn}{gene name to show}

\item{matSwitch}{matrix of gene inclusion time in different probabilistic cells projections}

\item{bins}{number of bins}

\item{par}{boolean. Use external plot parameters par() if TRUE. FALSE by default.}
}
\description{
Show cumulative probability of gene inclusion in correlated module
}
