% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{show.inclusion.summary}
\alias{show.inclusion.summary}
\title{Show summary cumulative probability of genes inclusion in correlated module}
\usage{
show.inclusion.summary(matSwitch, gns.show = rownames(matSwitch),
  bins = 8, par = FALSE)
}
\arguments{
\item{matSwitch}{matrix of gene inclusion time in different probabilistic cells projections}

\item{gns.show}{genes to show. By default all genes are shown}

\item{bins}{number of bins}

\item{par}{boolean. Use external plot parameters par() if TRUE. FALSE by default.}
}
\description{
Show summary cumulative probability of genes inclusion in correlated module
}
