% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{sig.explore}
\alias{sig.explore}
\title{Estimate optimal sigma parameter.}
\usage{
sig.explore(X, W = NA, M = as.integer(ncol(X)/2), n.sample = 1,
  sig.lims = seq(0.01, 0.2, 0.03), metrics = "cosine", p.power = 1,
  plot = TRUE, err.cut = 0.5, n.steps = 20, n.cores = 1)
}
\arguments{
\item{X}{gene (rows) vs cell (columns) expression matrix}

\item{M}{number of principal points in pptree modeling}

\item{n.sample}{number of sampled trees per each sigma}

\item{sig.lims}{a vector of sigma for which cross-validation estimated}

\item{metrics}{similarity measure. "cosine" or "euclidean"}
}
\value{
optimal sigma parameter
}
\description{
Using cross-validation criteria to select sigma parameter.
}
