% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{slide.cells}
\alias{slide.cells}
\title{Estimate probabil}
\usage{
slide.cells(r, root, leaves, wind = 50, mapping = TRUE,
  regions = NULL, n.cores = 10)
}
\arguments{
\item{n.samples}{a number of seed samplings.}

\item{seeds}{a vector of seeds to use. Overwrites n.samples.}
}
\value{
a list of pptree objects
}
\description{
Estimate probabil
}
