% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{slide.cors}
\alias{slide.cors}
\title{Assign cells in a probabilistic manner to non-intersecting windows along pseudotime}
\usage{
slide.cors(freq, mat, genesetA, genesetB)
}
\arguments{
\item{freq}{list of per-window cell probabilities. Outcome of slide.cells}

\item{mat}{expression matrix}

\item{genesetA}{a vector of fate-specific genes for first post-bifurcation branch}

\item{genesetB}{a vector of fate-specific genes for second post-bifurcation branch}
}
\value{
a list of local per-window matrices of average gene correlations with two fate-specific gene modules
}
\description{
Assign cells in a probabilistic manner to non-intersecting windows along pseudotime
}
