% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{synchro.path}
\alias{synchro.path}
\title{Estimate probabil}
\usage{
synchro.path(r, mat, genesetA, genesetB, w, step,
  n.mapping = length(r$cell.info), subtree, perm = FALSE)
}
\arguments{
\item{n.samples}{a number of seed samplings.}

\item{seeds}{a vector of seeds to use. Overwrites n.samples.}
}
\value{
a list of pptree objects
}
\description{
Estimate probabil
}
