% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{test.associated.genes}
\alias{test.associated.genes}
\title{Determine a set of genes significantly associated with the tree}
\usage{
test.associated.genes(r, X, n.map = 1, n.cores = 20, spline.df = 3,
  fdr.cut = 1e-04, A.cut = 1, st.cut = 0.8, summary = FALSE,
  subtree = NA, fdr.method = NULL, ...)
}
\arguments{
\item{r}{pptree object}

\item{X}{expressinon matrix of genes (row) vs cells (column)}

\item{fdr.cut}{FDR (Benjamini-Hochberg adjustment) cutoff on significance; significance if FDR < fdr.cut}

\item{A.cut}{cmplitude cutoff on significance; significance if A > A.cut}

\item{st.cut}{cutoff on stability (fraction of mappings with significant (fdr,A) pair) of association; significance, significance if A > A.cut}

\item{summary}{show plot of amplitude vs FDR of each gene's association. By default FALSE.}

\item{subtree}{restrict statistical assesment to a subtree}

\item{fdr.method}{a method to adjust for multiple testing. Default - Bonferroni. Alternatively, "BH" can be used.}
}
\value{
modified pptree object with a new field r$stat.association that includes pvalue, amplitude, fdr, stability and siginificane (TRUE/FALSE) of gene associations
}
\description{
Determine a set of genes significantly associated with the tree
}
