% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{visualise.clusters}
\alias{visualise.clusters}
\title{Visualize clusters of genes using heatmap and consensus tree-projected pattern.}
\usage{
visualise.clusters(r, emb, clust = NA, clust.n = 5, n.best = 4,
  best.method = "cor", cex.gene = 1, cex.cell = 0.1, cex.tree = 2,
  subtree = NA, reclust = FALSE)
}
\arguments{
\item{r}{pptree object}

\item{emb}{cells embedding}

\item{clust}{a vector of cluster numbers named by genes}

\item{n.best}{show n.best the most representative genes on the heatmap for each cluster}

\item{best.method}{use method to select the most representative genes. Current options: "pca" selects genes with the highest loading on pc1 component reconstructed using genes from a cluster, "cor" selects genes that have the highest average correlation with other genes from a cluster.}

\item{cex.gene}{size of gene names}

\item{cex.cell}{size of cells on embedding}

\item{cex.tree}{width of line of tree on embedding}

\item{subtree}{visualize clusters for a given subtree}

\item{reclust}{whether to reorder cells inside individual clusters on heatmap according to hierarchical clustering using Ward linkage and 1-Pearson as a distance between genes. By default is FALSE.}
}
\description{
Visualize clusters of genes using heatmap and consensus tree-projected pattern.
}
