% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{visualise.trajectory}
\alias{visualise.trajectory}
\title{Visualize branching trajectories of a particular gene.}
\usage{
visualise.trajectory(r, gene, X, cex.cell = 0.3, cex.lab = 2,
  cex.axis = 1.5, cex.main = 1, lwd.erbar = 0, lwd.t1 = 3,
  lwd.t2 = 0.2, switch.point = NA, subtree = NA)
}
\arguments{
\item{r}{pptree object}

\item{gene}{gene name}

\item{X}{matrix with a single row containing a gene expression levels (could be a vector of gene's expression). Columns of X reflect gene names.}

\item{cex.cell}{size of cells}

\item{cex.lab}{size of axis titles}

\item{cex.axis}{size of axis labels}

\item{cex.main}{size of title showing a gene name}

\item{lwd.erbar}{width of error bars for uncertainty of cell pseudotime assignment}

\item{lwd.t1}{width of the main branching trajectory}

\item{lwd.t2}{width of ensemble trajectories, typically thiner than that of main trajectory.}

\item{subtree}{visualise trajectory along a given subtree}
}
\description{
Visualize branching trajectories of a particular gene.
}
