% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{visualize.synchro}
\alias{visualize.synchro}
\title{Visualise local correlation trends of inter- and intra- module local correlations of fate-specific modules}
\usage{
visualize.synchro(crd)
}
\arguments{
\item{outcome}{of synchro function}
}
\value{
visualize combiation of ggplot2 objects
}
\description{
Visualise local correlation trends of inter- and intra- module local correlations of fate-specific modules
}
