% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{wcr}
\alias{wcr}
\title{calculate weighted correlation between columns of a matrix and a given vector}
\usage{
wcr(X, y, w)
}
\description{
calculate weighted correlation between columns of a matrix and a given vector
}
