% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{activation.statistics}
\alias{activation.statistics}
\title{Estimate optimum of expression and time of activation}
\usage{
activation.statistics(r, mat, root, leave, genes = rownames(mat),
  deriv.cutoff = 0.015, gamma = 1, n.mapping = 1, n.cores = 10)
}
\arguments{
\item{r}{ppt.tree object}

\item{mat}{expression matrix}
}
\value{
..
}
\description{
Estimate optimum of expression and time of activation
}
