% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestree.functions.R
\name{branch.specific.genes}
\alias{branch.specific.genes}
\title{Estimate optimum of expression and time of activation}
\usage{
branch.specific.genes(fork.de, stf.cut = 0.7, effect.b1 = 0.1,
  effect.b2 = 0.3, pd.a = 0, pd.p = 0.05)
}
\arguments{
\item{r}{ppt.tree object}

\item{mat}{expression matrix}
}
\value{
..
}
\description{
Estimate optimum of expression and time of activation
}
