% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{fig.cells}
\alias{fig.cells}
\title{Visualize cells assigned to each local window}
\usage{
fig.cells(emb, freq)
}
\arguments{
\item{emb}{embedding}

\item{freq}{cell probabilities of assignment to local windows.}
}
\value{
ggplot2 object
}
\description{
Visualize cells assigned to each local window
}
