% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{slide.cors}
\alias{slide.cors}
\title{Estimate probabil}
\usage{
slide.cors(freq, mat, genesetA, genesetB)
}
\arguments{
\item{n.samples}{a number of seed samplings.}

\item{seeds}{a vector of seeds to use. Overwrites n.samples.}
}
\value{
a list of pptree objects
}
\description{
Estimate probabil
}
