% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{synchro}
\alias{synchro}
\title{Estimate probabil}
\usage{
synchro(ppt, fpm, root, leaves, genesetA, genesetB, w, step, n.mapping,
  n.points = 100, span.smooth = 0.75, perm = FALSE)
}
\arguments{
\item{n.samples}{a number of seed samplings.}

\item{seeds}{a vector of seeds to use. Overwrites n.samples.}
}
\value{
a list of pptree objects
}
\description{
Estimate probabil
}
