% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.functions.R
\name{visualize.synchro}
\alias{visualize.synchro}
\title{Visualise local correlation trends of inter- and intra- module local correlations of fate-specific modules}
\usage{
visualize.synchro(crd, sc1 = NULL, sc2 = NULL, sc3 = NULL,
  tmax = NULL, vert = NULL)
}
\arguments{
\item{crd}{outcome of synchro function}

\item{sc1}{(2,and 3)specify ylims}

\item{tmax}{ymax (pseudotime)}

\item{vert}{add vertical lines at given pseudotimes}
}
\value{
visualize combiation of ggplot2 objects
}
\description{
Visualise local correlation trends of inter- and intra- module local correlations of fate-specific modules
}
